+++ changes.md
@@ -1,4 +1,7 @@
 #### 0.3.1
+* Added option to `skipTermCounting` when doing `wpp.editProperty`
+* Changed `wpp.scheduleStats` to returns counts of only published, private and future listings. 
+* Added `wpp.trashProperty` method for quicker removal.
 * Added [/wp-json/wp-rets-client/v1/systemPing] and [/wp-json/wp-rets-client/v1/getProperty] endpoints.
 * Added [/wp-json/wp-rets-client/v1/scheduleListings] endpoint and [wpp.scheduleListings] xml-rpc handler.
 * Added [/wp-json/wp-rets-client/v1/scheduleStats] endpoint and [wpp.scheduleStats] xml-rpc handler.